Add a new property to configure additional repos
When the `build.additional_maven_repo.mirror` is present,
The first configured repo for both configuration resolution
and buildscript will be `${build.additional_maven_repo.mirror}/mavenCentral`.
First Gradle Plugin resolution will be configured to
`${build.additional_maven_repo.mirror}/plugins`.

Note that all other repo configurations are left intact, so if this repo
does not work Gradle will fail over to the others.
If it does work, it's the one that will be used since it's the first.